Migrate to Python 3 and Ubuntu 24.04 (#1)
* start update to python3

* Create ci-ubuntu-24.yml

* test ci db

* try to fx workflow

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update setup.py

* BeautifulSoup workaround

* workflow fixes

* patch test

* try to use cassandra-driver

* Update ci-ubuntu-24.yml

* replace pycaptcha with captcha

* Revert "Update ci-ubuntu-24.yml"

This reverts commit a602650545123358670ba727db5d63bfc1567088.

* Update setup.py

* update workflow

* fixes

* fixes 2

* fixes 3

* simplejson

* Update simplejson.py

* Update simplejson.py

* Revert changes since 7cb64722a7cedffc2e64b490355703b876efa29b

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* Update ci-ubuntu-24.yml

* test

* test 2

* fixes 3

* Update setup.py

* Update app_globals.py

* update amqp

* Update setup.py

* Update __init__.py

* Create str.py

* Update filters.py

* Create _utils.py

* Create __init__.py

* Update baseplate_integration.py

* ColumnFamily compatibility shim

* ConsistencyLevel

* ActivityInfo stub

* Create sgm.py

* Update stats.py

* Update translation.py

* weberror shim

* baseplate

* Update app_globals.py

* Update __init__.py

* Update __init__.py

* Update __init__.py

* boto3

* Update zookeeper.py

* Update app_globals.py

* Update configparse.py

* Update app_globals.py

* Update db_manager.py

* sqlalchemy

* db_create_tables fix

* Update stats.py

* Update stats.py

* baseplate

* baseplate

* Create __init__.py

* Create observers.py

* Update server.py

* Create __init__.py

* Create test.ini

* Update test.ini

* test

* Delete test.ini

* Update tdb_lite.py

* test fixes

* test fixes 2

* Update tdb_cassandra.py

* Update cassandra_compat.py

* Update userrel.py

* Update admintools.py

* Update system_manager.py

* Update __init__.py

* Create wrapped.py

* Create batch.py

* cassandra

* Update sorts.py

* Update ci-ubuntu-24.yml

* Update wiki.py

* Update ci-ubuntu-24.yml

* types

* fixes

* Update admintools.py

* Update report.py

* Update bidding.py

* Update bidding.py

* Update bidding.py

* fixes

* fixes

* fixes

* Update mail_queue.py

* Update traffic.py

* Update traffic.py

* Update account.py

* Update middleware.py

* fixes

* fixes

* Update ci-ubuntu-24.yml

* Update __init__.py

* Update health.py

* fixes

* fixes

* fixes

* fixes

* Update _utils.py

* fixes

* Update comment_tree.py

* fixes

* Revert "fixes"

This reverts commit 1aae3fb0ff02020c8c9a7f9c3ca5cd1d42cbf2c4.

* fixes

* fixes

* fixes

* fixes

* Update wsgiapp.py

* Update base.py

* fixes

* Update __init__.py

* Update base.py

* Update loid.py

* fixes

* Revert "fixes"

This reverts commit ff0f7cef5e43e95e0dd4071d0cf8bfcd1cd3a6bf.

* Revert "Update loid.py"

This reverts commit c8f27ca814a925590643e40784d89d85700cd1dd.

* Reapply "Update loid.py"

This reverts commit cd8edb2b135b0ed653c134a5ba0b68bc33b15fa2.

* Fix Python 2 syntax for Python 3 compatibility

- Convert print statements to print() function calls in automoderator.py and hooks.py
- Replace ur"" and urrr"" string prefixes with r"" (Python 3 raw strings)



* Fix Python 2 string.lowercase in inject_test_data.py

Replace string.lowercase with string.ascii_lowercase for Python 3 compatibility.



* Update install scripts for Ubuntu 24.04 and Python 3.12

- Update repo URL from reddit/reddit to acalcutt/reddit
- Add support for Ubuntu 24.04 alongside legacy 14.04
- Update install_apt.sh with Python 3 packages for 24.04
- Change python to python3 in install scripts
- Fix Python 2 print statement to print()
- Update CI workflow to use verbose pytest output for debugging



* CI: Show full pytest traceback for debugging

Use --tb=long and pipe through head -500 to capture full error details.



* CI: Run only unit tests (skip functional tests for now)

Functional tests require full WSGI app setup which is hanging in CI.
Run only unit tests until the Pylons-to-Pyramid controller migration is complete.



* Update install scripts for Ubuntu 24.04 and use them in CI

- install_cassandra.sh: Add Ubuntu 24.04 support with Cassandra 4.x from Apache repo
- setup_cassandra.sh: Use cqlsh instead of pycassa for Python 3 compatibility
- install_services.sh: Update package names for Ubuntu 24.04, remove unavailable mcrouter
- ci-ubuntu-24.yml: Use install scripts instead of Docker for Cassandra and services



* Use Apache Cassandra repository instead of binary tarball

- install_cassandra.sh: Install Cassandra 4.1.x from debian.cassandra.apache.org
- Simpler installation using apt package manager
- cqlsh is included in the apt package



* Replace pkg_resources with importlib.metadata/resources

Addresses the pkg_resources deprecation warning by migrating to the
modern Python 3.9+ importlib APIs:

- plugin.py: Use importlib.metadata.entry_points() instead of
  pkg_resources.WorkingSet().iter_entry_points()
- app_globals.py: Use importlib.resources.files() instead of
  pkg_resources.resource_stream()
- providers/__init__.py: Use importlib.metadata.entry_points()
- tests/__init__.py: Use importlib.metadata.version() instead of
  pkg_resources.require()



* Update app_globals.py

---------

+++ scripts/migrate/backfill/msgtime_to_inbox_count.py
@@ -23,12 +23,11 @@
 
 import sys
 
+from pylons import app_globals as g
 from r2.lib.db import queries
 from r2.lib.db.operators import desc
 from r2.lib.utils import fetch_things2, progress
 from r2.models import Account, Message
-
-from pylons import app_globals as g
 
 
 def _keep(msg, account):
@@ -53,7 +52,7 @@ def _keep(msg, account):
 
     return True
 
-resume_id = long(sys.argv[1]) if len(sys.argv) > 1 else None
+resume_id = int(sys.argv[1]) if len(sys.argv) > 1 else None
 
 msg_accounts = Account._query(sort=desc("_date"), data=True)
